pyproject.toml
[tool.poetry]
name = "platinumshrimp"
version = "0.1.0"
description = ""
license = "MIT"
authors = ["Gustav Tiger <gustav@tiger.name>"]

readme = "README.md"

homepage = "https://github.com/Tigge/platinumshrimp"
repository = "https://github.com/Tigge/platinumshrimp.git"

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: POSIX :: Linux",
    "Topic :: Communications :: Chat :: Internet Relay Chat",
]

packages = [
    { include = "**/*.py" },
]

[tool.poetry.scripts]
platinumshrimp = "bot:main"

[tool.poetry.dependencies]
python = "^3.10"
python-dateutil = "^2.9.0"
requests = "^2.32.3"
feedparser = "^6.0.11"
irc = "^20.5.0"
pyzmq = "^26.2.0"

[tool.poetry.dev-dependencies]
requests-mock = "^1.12.1"
black = "^24.8.0"
rope = "^1.13.0"
coverage = "^7.6.1"

[build-system]
requires = ["poetry>=1.8.3"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 100
